WIP updation on existing features

+++ service/likeService.js
@@ -1,40 +1,52 @@
 const statusModel = require("../models/statusModel");
 const likeModel = require("../models/likeModel");
 const userModel = require("../models/userModel");
+const commentModel = require("../models/commentModel");
 
 async function addLike(req, res) {
   try {
     let id = req.params.id;
     let uid = req.id;
 
-    //check if the status is already liked by the user
-    let selectedStatus = await likeModel.find({
-      $and: [{ userId: uid }, { statusId: id }],
-    });
-
-    if (!selectedStatus.length) {
-      const newLike = {
-        statusId: id,
-        userId: uid,
-      };
-      let addedLike = await likeModel.create(newLike);
-
-      let status = await statusModel.findById(id);
-      status.totalLikes = status.totalLikes + 1;
-
-      await status.save();
-      res.status(200).json({
-        message: "Added new like !!",
-        data: addedLike,
+    //check if the status/comment exists
+    let content =
+      (await statusModel.findById(id)) || (await commentModel.findById(id));
+    if (content) {
+      //check if the status is already liked by the user
+      let selectedStatus = await likeModel.find({
+        $and: [{ userId: uid }, { statusId: id }],
       });
+
+      if (!selectedStatus.length) {
+        const newLike = {
+          statusId: id,
+          userId: uid,
+        };
+        let addedLike = await likeModel.create(newLike);
+
+        let status =
+          (await statusModel.findById(id)) || (await commentModel.findById(id));
+
+        status.totalLikes = status.totalLikes + 1;
+
+        await status.save();
+        res.status(200).json({
+          message: "Added new like !!",
+          data: addedLike,
+        });
+      } else {
+        res.status(501).json({
+          message: "Status/comment already liked by user !!",
+        });
+      }
     } else {
       res.status(501).json({
-        message: "Status already liked by user !!",
+        message: "enter correct status/comment id",
       });
     }
   } catch (error) {
     res.status(501).json({
-      message: "Failed to like the status",
+      message: "Failed to like the status/comment",
       error,
     });
   }
@@ -70,34 +82,44 @@ async function removeLike(req, res) {
     let id = req.params.id;
     let uid = req.id;
 
-    let selectedStatus = await likeModel.find({
-      $and: [{ userId: uid }, { statusId: id }],
-    });
-
-    //check if the status is already liked
-    if (selectedStatus.length) {
-      let removeLikeFromStatus = selectedStatus[0];
-      let removeLikeFromStatusId = removeLikeFromStatus._id.toString();
-      let unlikedStatus = await likeModel.findByIdAndDelete(
-        removeLikeFromStatusId
-      );
-
-      let status = await statusModel.findById(id);
-      status.totalLikes = status.totalLikes - 1;
-      await status.save();
-
-      res.status(200).json({
-        message: "Unliked the status",
-        data: unlikedStatus,
+    //check if the status/comment exists
+    let content =
+      (await statusModel.findById(id)) || (await commentModel.findById(id));
+    if (content) {
+      let selectedStatus = await likeModel.find({
+        $and: [{ userId: uid }, { statusId: id }],
       });
+
+      //check if the status is already liked
+      if (selectedStatus.length) {
+        let removeLikeFromStatus = selectedStatus[0];
+        let removeLikeFromStatusId = removeLikeFromStatus._id.toString();
+        let unlikedStatus = await likeModel.findByIdAndDelete(
+          removeLikeFromStatusId
+        );
+
+        let status =
+          (await statusModel.findById(id)) || (await commentModel.findById(id));
+        status.totalLikes = status.totalLikes - 1;
+        await status.save();
+
+        res.status(200).json({
+          message: "Unliked the status/comment",
+          data: unlikedStatus,
+        });
+      } else {
+        res.status(501).json({
+          message: "You can't unlike this status/comment",
+        });
+      }
     } else {
       res.status(501).json({
-        message: "You can't unlike this status",
+        message: "enter correct status/comment id",
       });
     }
   } catch (error) {
     res.status(501).json({
-      message: "Failed to unlike the status",
+      message: "Failed to unlike the status/comment",
       error,
     });
   }
@@ -112,7 +134,7 @@ async function statusLikedByUserId(req, res) {
     if (user) {
       let likedStatus = await likeModel.find({ userId: uid });
       res.status(200).json({
-        mesaage: "Liked status fetched successfully",
+        mesaage: "Liked status by user fetched successfully",
         data: likedStatus,
       });
     } else {
